Fix some type definition for android 5.1.0.

+++ src/com/android/reverse/smali/MemoryBackSmali.java
@@ -56,7 +56,7 @@ public class MemoryBackSmali {
 		return options;
 	}
 
-	public static boolean disassembleDexFile(int mCookie, String outDexName) {
+	public static boolean disassembleDexFile(long mCookie, String outDexName) {
 
 		long startTime = System.currentTimeMillis();
 		Logger.log("start disassemble the mCookie " + mCookie);
